Assignment Notification Fixes for 2.5.0
I took the agent assignment e-mail notification that x6iae created
and rebased it against release 2.5.0 (the version I am using). The
resulting code had a few minor test failures due to updates between
when x6iae did their work and the release of 2.5.0. This commit
addresses fixing those issues.

Bulk Mailing Fix
----------------

The first thing is updating the bulk mailing call. Previously it
passed the list of topics by just pulling from params but in
4716792f6b the system was updated so that the list of topic ids
is not always passed in. Instead other ways of loading the topics
might be used.

To compensate rather than read from the params I grab the ids from
the list of loaded topics.

History Fix
-----------

In ce173003 the templates were updated to include the full history
in the e-mail for reference. This is not used for the agent assign
notification (which is ok as AR relations are lazy). But since they
may have no history (initial topic being created) we get an error
as it tries to exclude the most recent post from the history.

By using the safe navigation operator I avoid the error.